### 2023-02-09 #5140 @Kevin-Mattheus-Moerman
@amiralih thanks for this submission. I am the AEiC for this track and here to help process initial steps and to check if this work is in scope for JOSS. I have just triggered a scope review by the editorial board, to check if this works conforms to our [substantial scholarly effort criteria]( https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The scope review will take about 2 weeks to complete.
