### 2023-02-23 #5140 @Kevin-Mattheus-Moerman
@amiralih the preliminary scope review determined this may be in scope for JOSS. I will remove the query-scope label now and look for a handling editor. Note that the handling editor and the reviewers may still have concerns on scope, but this initial review determined that in terms of functionality and extend of documentation/testing coverage this seems like it may be in scope. Thanks for your patience.
