+++ paper.bib
@@ -12,7 +12,10 @@
 @article{lessen2022molecular,
   title={Molecular mechanisms of spontaneous curvature and softening in complex lipid bilayer mixtures},
   author={Lessen, Henry J and Sapp, Kayla C and Beaven, Andrew H and Ashkar, Rana and Sodt, Alexander J},
-  journal={bioRxiv},
+  journal={Biophysical Journal},
+  volume={121},
+  number={17},
+  pages={3188--3199},
   year={2022},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Elsevier}
 }
