+++ paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={042413},
   year={2021},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevE.103.042413}
 }
 
 @article{lessen2022molecular,
@@ -17,5 +18,6 @@
   number={17},
   pages={3188--3199},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.bpj.2022.07.036}
 }
