+++ docs/src/fluctuation_analysis.md
@@ -8,3 +8,7 @@ fluctuation_spectrum
 area_expansion_modulus
 ```
 
+```@docs
+fluctuation_hq2_data
+```
+
