+++ v0.2.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"density_analysis.html#Density-Analysis","page":"Density Analysis","title":"Density Analysis","text":"","category":"section"},{"location":"density_analysis.html","page":"Density Analysis","title":"Density Analysis","text":"lipids_density_spectrum","category":"page"},{"location":"density_analysis.html#MembraneAnalysis.lipids_density_spectrum","page":"Density Analysis","title":"MembraneAnalysis.lipids_density_spectrum","text":"lipids_density_spectrum(;\n    pdb_file,\n    traj_file,\n    fs_file,\n    output_file,\n    lipids,\n    ref_atoms=Dict(lipid => lipid.ref_atom for lipid in lipids),\n    L_grid)\n\nCalculates lateral density (number per cell) spectrum of lipids in a lipid bilayer simulation trajectory and saves the results as a HDF5 file with labels XXXX_1 and XXXX_2 for lipid \"XXXX\" in the two leaflets.\n\nKeyword arguments\n\npdf_file: PDB structure file;\ntraj_file: trajectory file;\nfs_file: flucuation spectrum file;\noutput_file: output HDF5;\nlipids: a list of lipids of type Lipid as defined in lipids.jl;\nref_atoms: a dictionary of reference atoms for each lipid;\nL_grid: length of the lattice grid used to discretize the surface.\n\n\n\n\n\n","category":"function"},{"location":"density_analysis.html","page":"Density Analysis","title":"Density Analysis","text":"peptide_density_spectrum","category":"page"},{"location":"density_analysis.html#MembraneAnalysis.peptide_density_spectrum","page":"Density Analysis","title":"MembraneAnalysis.peptide_density_spectrum","text":"peptide_density_spectrum(;\n    pdb_file,\n    traj_file,\n    fs_file,\n    output_file,\n    lipids,\n    ref_residue,\n    L_grid)\n\nCalculates lateral density (number per cell) spectrum of the peptide using the CA atom of its reference residue and saves the results as a HDF5 file with labels PEP_1 and PEP_2 for the two leaflets.\n\nKeyword arguments\n\npdf_file: PDB structure file;\ntraj_file: trajectory file;\nfs_file: flucuation spectrum file;\noutput_file: output HDF5;\nlipids: a list of lipids of type Lipid as defined in lipids.jl;\nref_residue: residue number of the reference residue of the peptide;\nL_grid: length of the lattice grid used to discretize the surface.\n\n\n\n\n\n","category":"function"},{"location":"density_analysis.html","page":"Density Analysis","title":"Density Analysis","text":"lipids_radial_distribution","category":"page"},{"location":"density_analysis.html#MembraneAnalysis.lipids_radial_distribution","page":"Density Analysis","title":"MembraneAnalysis.lipids_radial_distribution","text":"lipids_radial_distribution(;\n    pdb_file,\n    traj_file,\n    fs_file,\n    output_file,\n    lipids,\n    ref_atoms=Dict(lipid => lipid.ref_atom for lipid in lipids),\n    lipid_A,\n    lipid_B,\n    max_r=20,\n    n_bins=200,\n    same_leaflet=true)\n\nCalculates the lateral (2-D) radial density function of lipid A and lipid B in a lipid bilayer simulation trajectory.\n\nKeyword arguments\n\npdf_file: PDB structure file;\ntraj_file: trajectory file;\nfs_file: flucuation spectrum file;\noutput_file: output file (density as a function of histogram bin center);\nlipids: a list of lipids of type Lipid as defined in lipids.jl;\nref_atoms: a dictionary of reference atoms for each lipid;\n
[…]
