### 2023-04-16 #5140 @majensen
@tonigi - thanks. The basic task is to consider the software package as submitted (found in its repository, listed at the top of this issue), installing as described in its documentation, and running it according to examples provided by the author in documents, or in another simple context as you like. A [checklist of items to review](https://joss.readthedocs.io/en/latest/review_checklist.html) will be created in the "review issue" (which will be generated when I have agreed reviewers), and you will use this to walk yourself through the review process. JOSS has [reviewer documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) that is helpful.
