### 2023-04-17 #5380 @editorialbot
[REVIEW]: MembraneAnalysis.jl: A Julia package for analyzing molecular dynamics simulations of lipid membranes
**Submitting author:** @amiralih (<a href="http://orcid.org/0000-0002-2580-3577">Amirali Hossein</a>)
**Repository:** https://github.com/amiralih/MembraneAnalysis.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
