### 2023-04-20 c6d8251
Added q_min for curvature (lipids)
