### 2023-04-22 #5380 @davidnsousa
Concerning functionality and documentation:

- Not all documented core functions have automated tests that confirm the functional claims.

- Tutorial is insufficient to understand the software usage: it does not provide example input data/files along with an explanation of how to structure input data; example data that allows to test every core function would be useful to confirm functional claims and tutor users along.

- The peptide_radial_distribution function lacks a documentation entry.
