### 2023-04-25 #5380 @tonigi
There are a couple of areas that could benefit from a bit more polish. First, the tutorial. I would recommend a self-standing working simple example with example data. It does not start with `using`, which means that one gets an error out of the box

    UndefVarError: DOPC_aa not defined

From what I can tell, it tries to open `s.pdb`, which is not in the repository. Could it be provided, and the tutorial made to analyze just one trajectory, for the sake of clarity?

In general, both the paper and the documentation are quite short. One could supplement the other, but in this case a lot of context is missing. For one, there is basically no mention of any other membrane analysis tools (state of the field), and there are many.
