+++ paper.bib
@@ -21,3 +21,75 @@
   publisher={Elsevier},
   doi={10.1016/j.bpj.2022.07.036}
 }
+
+@article{allen2009gridmat,
+  title={GridMAT-MD: a grid-based membrane analysis tool for use with molecular dynamics},
+  author={Allen, William J and Lemkul, Justin A and Bevan, David R},
+  journal={Journal of computational chemistry},
+  volume={30},
+  number={12},
+  pages={1952--1958},
+  year={2009},
+  publisher={Wiley Online Library},
+  doi={10.1002/jcc.21172}
+}
+
+@article{lukat2013apl,
+  title={APL@ Voro: a Voronoi-based membrane analysis tool for GROMACS trajectories},
+  author={Lukat, Gunther and Krüger, Jens and Sommer, Björn},
+  journal={Journal of chemical information and modeling},
+  volume={53},
+  number={11},
+  pages={2908--2925},
+  year={2013},
+  publisher={ACS Publications},
+  doi={10.1021/ci400172g}
+}
+
+@article{guixa2014membplugin,
+  title={MEMBPLUGIN: studying membrane complexity in VMD},
+  author={Guix{\`a}-Gonz{\'a}lez, Ramon and Rodriguez-Espigares, Ismael and Ram{\'\i}rez-Anguita, Juan Manuel and Carri{\'o}-Gaspar, Pau and Martinez-Seara, Hector and Giorgino, Toni and Selent, Jana},
+  journal={Bioinformatics},
+  volume={30},
+  number={10},
+  pages={1478--1480},
+  year={2014},
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btu037}
+}
+
+@article{buchoux2017fatslim,
+  title={FATSLiM: a fast and robust software to analyze MD simulations of membranes},
+  author={Buchoux, S{\'e}bastien},
+  journal={Bioinformatics},
+  volume={33},
+  number={1},
+  pages={133--134},
+  year={2017},
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btw563}
+}
+
+@article{bhatia2019memsurfer,
+  title={MemSurfer: A tool for robust computation and characterization of curved membranes},
+  author={Bhatia, Harsh and Ing{\'o}lfsson, Helgi I and Carpenter, Timothy S and Lightstone, Felice C and Bremer, Peer-Timo},
+  journal={Journal of chemical theory and computation},
+  volume={15},
+  number={11},
+  pages={6411--6421},
+  year={2019},
+  publisher={ACS Publications},
+  doi={10.1021/acs.jctc.9b00453}
+}
+
+@article{smith2021lipyphilic,
+  title={LiPyphilic: A Python toolkit for the analysis of lipid membrane simulations},
+  author={Smith, Paul and Lorenz, Christian D},
+  journal={Journal of chemical theory and computation},
+  volume={17},
+  number={9},
+  pages={5907--5919},
+  year={2021},
+  publisher={ACS Publications},
+  doi={10.1021/acs.jctc.1c00447}
+}
