+++ tutorial/tutorial.md
@@ -1,6 +1,6 @@
 # Example usage of `MembraneAnalysis.jl`
 
-In this tutorial we showcase some of the main functionality of the package by analyzing an atomistic simulations of a membrane containing DOPC lipids and %10 mol fraction cholesterol. `MembraneAnalysis.jl` needs a PDF structure file (here `s.pdb`) and trajectory files of any format supported by `Chemfiles.jl` (here `t1.nc` to `t200.nc`).
+In this tutorial we showcase some of the main functionality of the package by analyzing an atomistic simulations of a membrane containing DOPC lipids and %10 mol fraction cholesterol. `MembraneAnalysis.jl` needs a PDF structure file (here `s.pdb`) and trajectory files of any format supported by `Chemfiles.jl` (here `t1.nc` to `t100.nc`). The files needed for this tutorial can be found [here](https://doi.org/10.5281/zenodo.7908906).
 
 ## Setting analysis parameters
 
@@ -16,11 +16,13 @@ It is useful to define a set of parameters that we will need for input arguments
 - `output_dir`: a directory for all the output files generated by methods
 
 ```julia
+using MembraneAnalysis
+
 begin
   const pdb_file = "s.pdb"
   const traj_dir = "./"
   const traj_name = "t"
-  const traj_inds = 1:200
+  const traj_inds = 1:100
   const traj_ext = ".nc"
   const lipids = [DOPC_aa, Chol_aa]
   const L_grid = 15
