+++ docs/src/index.md
@@ -4,7 +4,7 @@ Documentation for `MembraneAnalysis.jl`, a package for analyzing molecular dynam
 
 !!! note
 
-    A tutorial demonstrating an example usage of the package can be found [here](https://github.com/amiralih/MembraneAnalysis.jl/blob/d72a933cb39a3beb5423224c4169a0f1494a5ec7/tutorial/tutorial.md).
+    A tutorial demonstrating an example usage of the package can be found [here](https://github.com/amiralih/MembraneAnalysis.jl/blob/main/tutorial/tutorial.md).
 
 
 ```@contents
