### 2023-05-15 b630080
New curvature spectrum, added interactions.
