+++ src/curvature_analysis.jl
@@ -350,7 +350,9 @@ end
         output_dir,
         lipids,
         ref_atoms=Dict(lipid => lipid.ref_atom for lipid in lipids),
-        q_max)
+        q_max,
+        q_min,
+        nqs)
 
 Calculates curvature spectrum of the lipids using their reference atom position. Assumes square (Lx = Ly) bilayer. Results for lipid "XXXX" will be stored in `XXXX_cqs.dat` in the output directory.
 
@@ -362,7 +364,11 @@ Calculates curvature spectrum of the lipids using their reference atom position.
 * `output_dir`: output directory;
 * `lipids`: a list of lipids of type `Lipid` as defined in `lipids.jl`;
 * `ref_atoms`: a dictionary of reference atoms for each lipid;
-* `q_max`: maximum q mode magnitude value to be used.
+* `q_max`: maximum q mode magnitude value to be used;
+* `q_min`: minimum q mode magnitude value to be used;
+* `nqs`: number of segments to divide the q range from q_min to q_max;
+* `tag_files`: a list of corresponding lipid interaction tag files of the trajectory files. Uses all lipids by default;
+* `tag_id`: the tag id of the selected lipids.
 
 """
 function lipids_curvature_spectrum(;
@@ -372,44 +378,61 @@ function lipids_curvature_spectrum(;
     output_dir,
     lipids,
     ref_atoms=Dict(lipid => lipid.ref_atom for lipid in lipids),
-    q_max
+    q_max=0.21,
+    q_min=0.01,
+    nqs=10,
+    tag_files=false,
+    tag_id=0
 )
     # finding indices of tail atoms
 
     atoms = PDBTools.readPDB(pdb_file)
-    leaflet_id = atom_leaflet(pdb_file=pdb_file, lipids=lipids)
 
     tail_atoms_inds = [a.index for a in atoms if (a.resname, a.name) in [(lipid.name, lipid.tail_atom) for lipid in lipids]]
 
-    # find |q| values (up to q_max) and corresponing indices
+    # find mid-segment |q| values (q_min to q_max)
+
+    dq = (q_max - q_min) / nqs
+    qs = collect((q_min + (dq / 2)):dq:q_max)
     
-    (L, _, _) = box_dimensions(traj_file=traj_files[1])
-
-    # including at least 9 q values (for N = 4) even if greater than q_max
-
-    N = max(Int(floor(q_max / (2π / L))), 4)
-    values, index_pairs = get_index_pairs(N)
-    qs = values * (2π / L)
- 
     # creating dictionary of vectors to record c_q values
     
     cqs = Dict()
 
     for lipid in lipids
         cqs[lipid.name] = Dict()
-        for q_i in 2:length(qs)
+        for q_i in 1:nqs
             cqs[lipid.name][q_i] = Float64[]
+        end
+    end
+    
+    # defining a flag for if there are tag files
+
+    tag_flag = true
+    if tag_files == false
+        tag_flag = false
+        tag_files = ["---" for fs_file in fs_files]
+    end
+
+    # create index to resnum dictionary
+
+    index_to_resnum = Dict()
+
+    for a in atoms
+        if a.resname in [lipid.name for lipid in lipids]
+            index_to_resnum[a.index] = a.resnum
         end
     end
 
     # reading trajectories    
 
-    for (traj_file, fs_file) in zip(traj_files, fs_files)
+    for (traj_file, fs_file, tag_file) in zip(traj_files, fs_files, tag_files)
         traj = Chemfiles.Trajectory(traj_file)
         n_frames = Int(Chemfiles.size(traj))
         
         hq = h5read(fs_file, "hq")
         l_id = h5read(fs_file, "l_id")
[…]
