+++ docs/src/interactions.md
@@ -1,0 +1,6 @@
+# Interactions
+
+```@docs
+count_interactions
+```
+
