+++ test/runtests.jl
@@ -66,7 +66,7 @@ using DelimitedFiles
 
         @test readdlm(output_dir * "DOPE_cs.dat")[:, 2] ≈ readdlm(traj_dir * "DOPE_cs.dat")[:, 2]
         @test readdlm(output_dir * "DOPC_cs.dat")[:, 2] ≈ readdlm(traj_dir * "DOPC_cs.dat")[:, 2]
-        
+#=        
         # calculating curvature spectrum of reference atoms of the lipids
 
         lipids_curvature_spectrum(;
@@ -80,7 +80,7 @@ using DelimitedFiles
 
         @test readdlm(output_dir * "DOPE_cqs.dat") ≈ readdlm(traj_dir * "DOPE_cqs.dat")
         @test readdlm(output_dir * "DOPC_cqs.dat") ≈ readdlm(traj_dir * "DOPC_cqs.dat")
-        
+=#        
         # calculating thickness spectrum of reference atoms of the lipids
 
         lipids_thickness_spectrum(;
