+++ test/runtests.jl
@@ -3,7 +3,7 @@ using Test
 using DelimitedFiles
 
 @testset "MembraneAnalysis.jl" begin
-        
+
         # setting system parameters
 
         begin
@@ -17,6 +17,7 @@ using DelimitedFiles
             if !isdir(output_dir) mkdir(output_dir) end
             fs_file = output_dir * "fs_1.h5"
             ds_file = output_dir * "ds_1.h5"
+            tag_file = output_dir * "tag_1.h5"
         end
 
         # calculating fluctuation spectrum of trajectories
@@ -39,7 +40,38 @@ using DelimitedFiles
         )
 
         @test readdlm(output_dir * "KA.dat") ≈ readdlm(traj_dir * "KA.dat")
+        
+        # calculating box dimensions and area
 
+        box_dimensions(;
+            traj_file=traj_file,
+            area_file=output_dir * "A.dat",
+            box_dims_file=output_dir * "box_dims.dat")
+        
+        @test readdlm(output_dir * "A.dat")[1] ≈ readdlm(traj_dir * "A.dat")[1]
+        @test readdlm(output_dir * "box_dims.dat")[1:3] ≈ readdlm(traj_dir * "box_dims.dat")[1:3]
+
+        # calculating |hq|^2 vs |q| data from fluctuation spectrum
+        
+        Lx, Ly = readdlm(output_dir * "box_dims.dat")[1:2]
+
+        fluctuation_hq2_data(;
+            box_dims=(Lx, Ly),
+            fs_files=[fs_file],
+            output_dir=output_dir,
+            q_max=q_max)
+        
+        @test readdlm(output_dir * "hq2.dat")[:, 2] ≈ readdlm(traj_dir * "hq2.dat")[:, 2]
+        
+        # calculating blocking error
+        
+        @test blocking_error(sin.(1:1000)) ≈ 0.0012181125344636647
+        
+        # calculating index pairs
+        
+        @test get_index_pairs(2)[1] ≈ [0.0, 1.0, 1.4142135623730951, 2.0]
+        @test get_index_pairs(2)[2] == [[(0, 0)], [(0, 1), (1, 0)], [(1, 1)], [(0, 2), (2, 0)]]
+        
         # calculating mean height of lipids heavy atoms from the midplane
 
         lipids_atoms_height(;
@@ -66,7 +98,28 @@ using DelimitedFiles
 
         @test readdlm(output_dir * "DOPE_cs.dat")[:, 2] ≈ readdlm(traj_dir * "DOPE_cs.dat")[:, 2]
         @test readdlm(output_dir * "DOPC_cs.dat")[:, 2] ≈ readdlm(traj_dir * "DOPC_cs.dat")[:, 2]
-#=        
+
+        # calculating reference atoms
+
+        ref_atoms = find_ref_atoms(;
+            input_dir=traj_dir,
+            lipids=lipids,
+            z=5,
+        )
+
+        @test ref_atoms == Dict(DOPE_m2 => "C3A", DOPC_m2 => "C3A")
+
+        # calculating voronoi shells
+
+        v_shells = voronoi_shells(;
+           points=hcat([[i, j] for i in 1:2:9, j in 1:2:9]...),
+           box_dims=(10, 10),
+           center=[4, 4],
+           n_shells=2
+        )
+
+        @test v_shells == Dict(2 => [1, 2, 3, 4, 6, 9, 11, 14, 16, 17, 18, 19], 1 => [7, 8, 12, 13])
+        
         # calculating curvature spectrum of reference atoms of the lipids
 
         lipids_curvature_spectrum(;
@@ -75,12 +128,33 @@ using DelimitedFiles
             fs_files=[fs_file],
             output_dir=output_dir,
             lipids=lipids,
-            q_max=q_max
-        )
[…]
