### 2023-05-24 #5380 @amiralih
@majensen I have tagged version 0.3.0 that addresses the issues raised by the reviewers:
- The paper has been updated to emphasize key functionality of the package that distinguishes it from existing alternatives.
- The tutorial has been updated with a set of accompanying trajectory files to be self-sufficient.
- Test coverage has been expanded.
