### 2023-06-03 #5380 @tonigi
Trying to run the tutorial I get

```
julia> for i in traj_inds
           traj_file = traj_dir * traj_name * string(i) * traj_ext
           output_file = output_dir * "fs_$(i).h5"

           fluctuation_spectrum(;
               pdb_file=pdb_file,
               traj_file=traj_file,
               output_file=output_file,
               lipids=lipids,
               L_grid=L_grid
           )
       end
ERROR: SystemError: opening file "s.pdb": No such file or directory
```

Where is `s.pdb`?
