### 2023-06-03 #5380 @amiralih
@tonigi The file `s.pdb` is available (along with the trajectory files) at the link provided in the tutorial:
https://zenodo.org/record/7908906
