### 2023-06-03 #5380 @davidnsousa
Regarding the improvements I have updated the checklist above.

- A statement of need is still missing in the documentation.
- A description of how to structure/generate input data would improve the documentation to a more satisfactory level.
- The functional claims cannot yet be confirmed from the tutorial.

I tested the tutorial with the example files from the link provided: `s.pdb` and two trajectory files `t1.nc` and `t2.nc.

Calculating mean sampled curvature of heavy atoms of the lipids fails with:

`ERROR: LoadError: UndefVarError: 'traj_id' not defined`

I have corrected `traj_id` to `traj_inds` which solves the problem. The tutorial should be corrected.

Transverse curvature bias analysis fails with:

`ERROR: LoadError: ArgumentError: Cannot open './A.dat': not a file`
