### 2023-06-04 #5380 @tonigi
Uhm... isn't >14 GB of files excessive for a tutorial? Also considering that there doesn't seems  to be a way to download all the 100 files at once from Zenodo.
