### 2023-06-05 #5380 @amiralih
@majensen @tonigi @davidnsousa I have updated the tutorial to link to one zipped folder which can be downloaded at once and contains 10 trajectories (~600 MB in total). I have also added a sentence clarifying the need to unzip the folder and navigate to it in order to run the tutorial as is. I have also fixed the two mistakes in the tutorial which were causing errors. The statement of need of the paper has also been added to the first page of the documentation.
