### 2023-06-06 24fd76b
Added input files guide to docs.
