### 2023-06-06 bb9fb14
Added input files guide to docs.
