### 2023-06-06 #5380 @amiralih
@davidnsousa Thanks for the insight, I have added a section to the first page of documentation that provides information about organizing the input files and files containing generated intermediate values.
