### 2023-06-07 #5380 @davidnsousa
- The tutorial runs successfully and automated tests are run that confirm the main functionality of the package.
- The documentation has improved to a satisfactory level.

In sum, all issues I have identified have been addressed. I have updated the checklist above.

From my side, the paper and package are ready and acceptable for publication :)
