### 2023-06-26 #5380 @majensen
@amiralih Have a look at [this small tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) for creating an open archive. If you can do this and report the DOI back here, I can set that and I believe we will be close to completion.
