### 2023-07-04 #5380 @majensen
Hi @amiralih - can you make the following minor changes in the paper:
l.7 - surrounding -> surroundings
l.24 - Nuetral -> Neutral
l.45 - no comma after “and”
l. 70- capitalize Voronoi
thanks!
