+++ paper.bib
@@ -35,7 +35,7 @@
 }
 
 @article{lukat2013apl,
-  title={APL@ Voro: a Voronoi-based membrane analysis tool for GROMACS trajectories},
+  title={APL@ Voro: a {Voronoi}-based membrane analysis tool for GROMACS trajectories},
   author={Lukat, Gunther and Krüger, Jens and Sommer, Björn},
   journal={Journal of chemical information and modeling},
   volume={53},
