+++ CITATION.cff
@@ -1,0 +1,34 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Hossein
+  given-names: Amirali
+  orcid: "https://orcid.org/0000-0002-2580-3577"
+- family-names: Sodt
+  given-names: Alexander J.
+  orcid: "https://orcid.org/0000-0002-5570-8212"
+doi: 10.5281/zenodo.8084178
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Hossein
+    given-names: Amirali
+    orcid: "https://orcid.org/0000-0002-2580-3577"
+  - family-names: Sodt
+    given-names: Alexander J.
+    orcid: "https://orcid.org/0000-0002-5570-8212"
+  date-published: 2023-07-06
+  doi: 10.21105/joss.05380
+  issn: 2475-9066
+  issue: 87
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5380
+  title: "MembraneAnalysis.jl: A Julia package for analyzing molecular
+    dynamics simulations of lipid membranes"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05380"
+  volume: 8
+title: "MembraneAnalysis.jl: A Julia package for analyzing molecular
+  dynamics simulations of lipid membranes"
